requirements.txt
requests
numpy
torch
torchvision
Pillow
opencv-python
matplotlib
imutils
dlib
PyQt5